Callback no worky
I tried to make a system for custom callback functions for each charachteristic but it's giving me trouble and I'm not sure it's so neccisary since I'll be responding to the callbacks in the loop anyway. I'm gonna restructure but I wanted to push to have a copy of what I did.